Fixed response issues
Fixed some issues when asking for user response to continue. Before if
something other than yes or no was entered the program would either
continue searching or end. Now if something else other than yes or no is
entered, the program informs the user and asks for them to respond
again.